test(email): consolidate isResend cases into existing spec
