+++ backend/src/lib/email/templates/network-invitation.template.spec.ts
@@ -50,4 +50,51 @@ describe('networkInvitationTemplate', () => {
     });
     expect(out.subject).not.toMatch(/[\r\n]/);
   });
+
+  test('renders the original invitation when isResend is omitted', () => {
+    const out = networkInvitationTemplate({
+      networkName: 'Experiment X',
+      apiKey: 'a-very-secret-key',
+      connectCommand: 'openclaw connect --key=a-very-secret-key',
+    });
+    expect(out.subject).toBe('Your invitation to Experiment X');
+    expect(out.text).toContain("You've been added to Experiment X");
+    expect(out.text).not.toContain('previous key has been revoked');
+    expect(out.html).not.toContain('previous key has been revoked');
+  });
+
+  test('renders the original invitation when isResend is false', () => {
+    const out = networkInvitationTemplate({
+      networkName: 'Experiment X',
+      apiKey: 'a-very-secret-key',
+      connectCommand: 'openclaw connect --key=a-very-secret-key',
+      isResend: false,
+    });
+    expect(out.subject).toBe('Your invitation to Experiment X');
+    expect(out.text).not.toContain('previous key has been revoked');
+  });
+
+  test('renders the refreshed variant when isResend is true', () => {
+    const out = networkInvitationTemplate({
+      networkName: 'Experiment X',
+      apiKey: 'a-very-secret-key',
+      connectCommand: 'openclaw connect --key=a-very-secret-key',
+      isResend: true,
+    });
+    expect(out.subject).toBe('Your access key for Experiment X (refreshed)');
+    expect(out.text.startsWith('Your previous key has been revoked. Use the key below going forward.')).toBe(true);
+    expect(out.html).toContain('Your previous key has been revoked. Use the key below going forward.');
+    expect(out.text).toContain('a-very-secret-key');
+    expect(out.text).toContain('openclaw connect --key=a-very-secret-key');
+  });
+
+  test('strips control chars from refreshed subject just like the original', () => {
+    const out = networkInvitationTemplate({
+      networkName: 'Bad\r\nNetwork',
+      apiKey: 'a-very-secret-key',
+      connectCommand: 'openclaw connect --key=a-very-secret-key',
+      isResend: true,
+    });
+    expect(out.subject).toBe('Your access key for Bad Network (refreshed)');
+  });
 });
